pyproject.toml
[tool.poetry]
name = "to_do_django_factory"
version = "0.1.0"
description = ""
authors = ["Besnard Consulting <contact@besnard-consulting.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.1"
black = "^20.8b1"
mypy = "^0.782"
flake8 = "^3.8.4"
pytest-django = "^3.10.0"
factory_boy = "^3.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
